Add autorun bootscript support for u-boot
Loading the bootloader from SD is horribly slow, this allows you to
get the bootloader from the QSPI flash, and then optionally get the
rest from SD card or do whatever you like in the autorun script.

It is also possible to create scripts to flash the image on SD to
somewhere else, and lots of other options.