Add UBIFS support and a tiny image to showcase it, boot enhancements
Target is to squeeze it all into the QSPI flash. Enable the UBIFS in
the kernel, repartition the flash to be able to use UBI, add an
image that's small enough to fit, and figure out the correct
settings for mkubifs and friends.

Also make the QSPI and NOR boot smarter. u-boot will copy the image directly
from the memory-mapped flash address to the load address in RAM. Since the
size of the image is in the header, this reduces the number of bytes to be
transferred considerably (from 4MB to 3MB currently)