linux-adi: Build framebuffer related components as modules
Build everything that relates to FPGA logic as a module. This allows
the system to boot first and load the logic later on from the rootfs.
It also makes it possible to chose between implementations and what
logic to load into the FPGA.

To start a framebuffer, first program the FPGA, then load the modules
in this order:
i2c-xiic xilinx_dma snd-soc-adv7511-hdmi adv7511 analogdrm