Add an FPGA bit image converter and loader
The script converts an FPGA "bit" file into a "bin" file suitable
for programming the Zynq FPGA. It adds an initscript to the rootfs
that loads the bin file into the FPGA at boot.
The bit-to-bin conversion is done on the host system, and can be used
for virtually any FPGA, not just the Zynq. For a Spartan you will need
to omit the --flip switch that reverts the 32-bit byte ordering.