zedboard: Add wifi support (as modules) and OLED driver (as module)
OLED driver loads but doesn't work, so it probably needs something in
the PL to connect the GPIO lines to the actual device.

Build modules for wifi support. Added a few common USB sticks. To
use wifi in your image, you'll need to add the firmware, iw and
wpa-supplicant packages.