Add support for systemd and a sample configuration
Add CONFIG_CGROUPS to the kernel config. This is required for systemd
to function correctly.

Add a tiny-systemd.conf distro. Note that I have not gotten this to
actually work yet.

Reduce the DISTRO_FEATURES set so that less packages get built, as
this image is expected to run headless.