zynq-zc702 kernel: Add I2C devices (eeprom, keypad, oscilator)
Create devicetree entries as in the Xilinx tree, and activate the
corresponding drivers in the kernel. This adds the 1k eeprom chip,
the si570 oscilator, the tca6416 gpio expander, and the RTC.

Activating the ucd9248 power management chips leads to lots of I2C
errors "i2c i2c-0: sendbytes: NAK bailout." so I've commented them
out of the device tree.