fpga-image: Build the logic using XPS
Instead of requiring a binary image, use the XPS program to build the
bitfile. This will take an hour or so, but at least you won't have to
to it all by hand, and it happens only once.

To make this work, you'll need to install the Xilinx tools (v14.4
recommended, not tested with others) and probably set the XILINX_
variables in your local.conf (the defaults should be okay if you
installed them in their default locations at /opt/xilinx).

Make sure there is no + (plus sign) in your TMPDIR somewhere, the
XPS tool will fail in very vague ways when that happens. I've worked
around that limitation by removing all "+" characters from the
version string, so that the WORKDIR doesn't add any plusses.