fpga-image: Add SDK export, use proper path locations
Export the hardware SDK into the sysroot (/usr/share/xilinx_sdk)
This allows other packages to gather information from the hardware
design, like the bitstream, memory mappings and code for the FSBL.

Replace hard-coded "/etc" with "${sysconfdir}". Move fpga.bin to
the ${datadir} with the other binaries. Add ".sh" to the initscript
filename so it will be sourced (quicker) instead of forked.

Also cleaned the SRCPV calculation, and remove the dependency
on the compiler toolchain. Since this step uses mostly one core,
it'll help to start it ASAP in the build.