linux-milo: Use the layered devicetrees in the kernel
Commit 6bb0fd7a3c1b2022cd440c005c094f33e1f7e400 introduced a nice layered
approach for the device trees. Embrace the idea by referring to these dts
files instead of placing it in the OE layer.

Note that the resulting binary devicetree will no longer be called
uImage-devicetree.dtb, but it'll come out as "uImage-zynq-zed-adv7511.dtb"
or something similar, usign the basename of KERNEL_DEVICETREE in the
recipe.

Update to the latest kernel which has the additions to the devicetrees to
make this image work correctly.