Migrate design to latest version for Zedboard
This (probably) breaks the zc702 build. I'm very sorry for that.

New zedboard design that uses PL330 DMA for audio.
Upgrade to (merged) kernel 3.9 (from ADI and XILINX)
Update defconfig and devicetree to match the new configuration.

The FSBL must be built outside the tree for now. u-boot will fail
to build unless you change the recipe to fetch a proper fsbl.elf.