linux-milo: Allow kernel to be upgraded
Install the kernel image in /tmp/boot.
Add a postinstall to the kernel, when booting from flash program
the kernel into mtd3, when booting from SD write the kernel to the
FAT partition on the SD card.

This allows a new kernel to be installed using "opkg upgrade".

TODO: Upgrade the devicetree as well.