Allow image-dependent fpga-image (bitstream) generation
Remove "fpga-image" from machine dependencies, the dependency is not correct. This
means that images must somehow include the dependency by themselves.
Remove kernel-modules from the machine dependency list, not all images need it.

Split fpga-image into an include fiel with common code, and a fpga-image-adi that
fetches and builds the ADI reference design.
Put the initscript loader into a separate package to remove extra file dependencies and
code duplication in fpga-image implementations.
Convert the fpga-bit-to-bin into a native package so that it ends up in the sysroot and
recipes like fpga-image can DEPEND on it and use it without copying files around.