Integrate new ADI design and upgrade to ISE 14.6
Bring the FPGA design up to date. This also requires a version 14.6 of
the ISE tools and won't compile with 14.4. Note that the previous design
did not compile in 14.6 at all.

Changes in the video DMA cores are in the xcomm_zynq_new_pcore_regmap
kernel branch, so linux-milo also had to switch to that branch in order
to get working HDMI output.