Add udhcpd scripts for using USB in gadget ethernet mode and for wifi AP
Busybox has a built-in DCHP server. These scripts enable its use
for usb and wifi access point usage. Include them on a board, and
modify /etc/network/interfaces (from init-ifupdown) to assign static
addresses to the usb0 or wlan1 interfaces, e.g.:

auto usb0
iface usb0 inet static
        address 192.168.232.1
        netmask 255.255.255.240

With the USB in gadget/periferal or OTG mode, this will automatically
assign an address to the host side, so you can just plug it into any
Windows or Linux PC and use the USB link as if it were network.