Add fpga-image-wizard recipe for pre-built bitstreams
If you have Vivado and just want to merge the generated files into the
image (the wizard wants to do so), this recipe allows to do that by
simply copying the bitstream files instead of having them built.

For this recipe to work, define PATH_TO_BITSTREAMS and PATH_TO_HW_EXPORT
in your configuration file so it points to the output files.