layer.conf: don't let META_ZYNQ_BASE value end up in sstate-cache hashes
When compiling on different machines, META_ZYNQ_BASE is likely to be different.
This value ends up in the sstate because it's sometimes used for the license
file location. Force its vardepvalue to be something static to prevent this.
The bad news is that the build will now consider your FPGA image out of date.
The good news is that sstate-cache might now actually work for FPGA images.