busybox/defconfig: Remove -R option from dhcpc commandline
The "-R" option sends a "release" to the DHCP server when shutting
down your box. This may be friendly towards open access points and
similar situations, but this box is likely to stay put, so giving
up the IP address only leads to losing your IP on a proper shutdown
and reboot cycle.