Allow topic-miami boards to act as ethernet gadgets
When the gadget modules have been built in the kernel, they will probe before
the USB EHCI driver starts, because the EHCI has a dependency on the USB reset
line which is wired through a I2C GPIO controller.
To fix this, build the g_ether driver as module so it will probe later. Also
offers multiple roles for the device, by loading the desired mode.

Fix the configuration for udhcpd to have the correct IP address.

The USB must be in gadget or OTG mode for this to work.