Revised MACHINE naming
To support combinations of carrier boards and FPGA families, make the
MACHINE variable a compound of "<SOM>-<CARRIER>-<FPGA>" and split
this up into the MACHINE_ARCH, SOM_FAMILY and FPGA_FAMILY variables.

Thus by default, this yields just one kernel and bootloader for both
the 7015 and 7030 types. The FPGA bitstream on the other hand will only
depend on the family of the hardware and no longer on the board.

This way, 3 carriers and 2 SOMs combine into 6 machine combinations but
only use a single configuration file.

Provide MACHINE_EXTRA_RRECOMMENDS to load recommended modules for each
board.