miami-xilinx-sdk: Reduce VDMA clocks from 150 to 100MHz
Experiments show that the minimum clock is way below 100MHz, the VDMA
retrieves 2 pixels per clock, for a single full-hd stream it would
need only 60MHz. To support two displays, and to allow for extra
headroom, 100MHz is a good balance.

The VDMA clock is driven by FCLK1. It could be routed to FCLK0 to
accomplish the same effect (and save on clock domains).