Add support for FPGA recipes that only build partial bitstreams
fpga-image-partials is mostly a copy of fpga-image.inc, some refactoring
is imminent.

The recipe expects a "generate_partial.sh" script in the repo that
does the heavy lifting of asking Vivado to load the given checkpoint
and build the partial streams into the "bitstreams" subdir.

The "static" recipe is expected to export its checkpoint and optionally
some IPs to the sysroot, these locations are passed in environment vars.