Kernel: Get rid of L2C warning message, and intern backlight
Added kernel commit that gets rid of the message:
  L2C: platform provided aux values match the hardware, so have no effect...

Move the backlight module into the kernel, since it doesn't require
logic to be programmed, and shave a few ms of the boottime. The module
loading was incorrect in vdmafb, due to a typo, but it got loaded
automatically. Since backlight is a vdmafb dependency, it should not
have been in the machine install list.