Add support for topic-miami-florida-pci carrier boards
The topic-miami-florida-pci board is a PCIe card that fits into a PCIe GEN2 X4
slot of a PC, allowing access to the FPGA through the PCIe interface.
The board also contains USB OTG, Gigabit Ethernet, SD card and the multi-serial
consolse, usign the same components as the "gen" board. All connectors except
JTAG are accesible through the back bracket. There are also a few LEDs that
can be controlled through logic.