udhcpd-iface-config: Store leases file in separate RAM locations
Stores the "leases" file in /var/run/ which is usually in RAM. The server
isn't intended to live very long, it's just for assigning addresses to
USB connections and such. No need to write it to the filesystem. Also
split them into separate files per interface since multiple instances
may be running.

Fixes the problem where my host managed to get a ".3" adress and it was
pretty hard to get it to .2 again.