conf/topic-miami: Remove link between FPGA_FAMILY and MACHINE
FPGA_FAMILY is related to MACHINE, but when set with "=", sstate-cache
will remember the dependency and rebuild packages with FPGA_FAMILY as
architecture needlessly.

When using ":=" to assign the value, this connection is removed and so
sstate-cache can be properly shared between machines.

Unfortunately, sstate-cache also detects this, and so this commit will
cause all FPGA images to be rebuilt.