my-image: Add udhcpd-iface-config to better support USB gadget mode
Connecting a PC to the USB port of a Florida board sets up a USB network link
in hte default configuration. The udhcpd-iface-config adds a DHCP server to
the system so that the PC automatically configures this interface, and the
link is actually immediately available without further need of commands.

The package was already in the feed, but it wasn't added to the standard
image, so the USB link didn't work out of the box.