topic-miami: Use "bitbang" I2C controller instead of the Cadence IP
Excerpt from the corresponding kernel commit:
The "hard" I2C controller in the Zynq appears to be plagued with bugs that
won't be fixed in hardware, and they cannot all be worked around. Symptom
on the Florida boards is that it is hard to get the adv7611 HDMI input chip
to work reliably with this controller. When replaced with the "bitbang"
controller on the same MIO pins, things do work okay.