Remove "mp3-image.bb", it uses obsolete gstreamer 0.10
gstreamer 0.10 is deprecated and has been removed from OE. The image recipe
depends on it, so best remove it to prevent people trying to build it.

Suggesting to use gstreamer for MP3 playback isn't good advise, if you want
an image with MP3 playback capability, just add "mpg123" to it.