Move to u-boot 2016.07, adjust QSPI partitioning, remove cadence I2C
U-boot 2016.07 is Xilinx tag 2016.2

This makes the SPL part (just) exceed 64k, so we expand the QSPI
partitions and make them equal to the partitioning of the "plus".

linux-topic: Adjust kernel for u-boot changes

New QSPI partitioning scheme, same for plus and regular boards.

Get rid of the faulty I2C controller in the Zynq, use GPIO driver.
This solves the 'missing' voltage regulator and IO expander.

Update the kernel to the xilinx/master state.