templates: Do not pick a default machine
Having the "med" board as a default was okay when we only had two boards
that were 99% equal. Now that there's a larger number of carriers to
choose from, it's better to not have a default at all, and trigger the
end user to make a selection that actually matches what's on his desk.