create-boot-archive.sh: Add script to create boot archive
Filling the boot partition requires renaming some of the files. This
script creates a tar file of the boot partition contents which can
be extracted to a FAT partition. Set MACHINE and DTB environments
before calling the script. The resulting tar is written to:
tmp-glibc/deploy/images/${MACHINE}/boot.tar.gz