u-boot: Move to mainline, detect RAM size at boot
Move to mainline u-boot. To get the integrated support for the Miami board, the
SRCREV must be set to a revision beyond 2016.11 so set that manually.

Add a patch to detect the DDR RAM size at boot. The memory size on the Miami is
a customer choice and can be 256, 512 or 1024MB. The timings are compatible, so
the boards can share the same bootloader if they just detect what the installed
memory size is.