kernel: Workaround for "stale" defconfig being used in build
In short, when I make a change to the "defconfig" file, the kernel is
rebuilt (which is correct), but the resulting kernel has been built using
the old defconfig from a previous build, instead of the new one.

What seems to be the problem is this code in kernel.bbclass:

    # Copy defconfig to .config if .config does not exist. This allows
    # recipes to manage the .config themselves in do_configure_prepend().
    if [ -f "${WORKDIR}/defconfig" ] && [ ! -f "${B}/.config" ]; then
        cp "${WORKDIR}/defconfig" "${B}/.config"
    fi

This keeps any existing ".config" file if it happens to still be in the $B
path, which is the case if you're rebuilding a kernel.

The workaround is to forcibly execute that step, so that "stale" configs
will not survive.

This is a workaround, pending a real solution in OE-core.