kernel: Make ADV7511 an internal module
If the adv7511 probes after the hdmi_tx driver, bad things appear to happen.
With systemd, the module load order is different, causing a crash because of
this.

The adv7511 is a small I2C driver, so it can be built into the kernel so it
probes early, and that solves the issue.