fpga-image: Support Linux FPGA manager as firmware loader
This adds support for loading fpga.bin using the FPGA manager in Linux. On the ZynqMP this
is the only waty to load bitstreams, on the Zynq this can be used as an alternative. Use
the "fpgamgr" machine feature to select which is to be integrated.

This also introduces FPGA_BITSTREAM_PATH and FPGA_STATIC_BITSTREAM_PATH variables that
can be set in recipes. This will also trigger a rebuild of all FPGA image unfortunately
because OE considers a dependency tree change a build trigger even if the resulting variables
retain the original values.