machine/topic-miami: Remove RSI WiFi drivers
On the florida v3r1 the chip has been removed in favor of an extra PMOD.
On the older board the chip was there but didn't actually work reliably,
lack of support from Redpine.
Remove the WiFi kernel modules from the recommended-to-install list and
don't build them. Also remove the fpga-power module which was only used
by sdhc1 for this wifi chip.