topic-miamimp: Build PMU firmware from source code
Remove the prebuilt pmufw.bin blob from the repo.

Build a pmu-firmware version that has the required configuration built-in. The FSBL
sent this to the PMU, but the license on the config code prohibits us from doing so
in u-boot SPL. Work around this by building the config object into the PMU itself.