kernel: Store ethernet MAC addresses in EEPROM on SOM
Implemented an nvmem implementation for mainline linux, backported it
and implemented it for the macb driver. This stores the MAC address in
the EEPROM on the Miami SOM so it can be made permanent. Writing an
address into EEPROM is a manual action still, write at offsets 0x1e8
and 0x1ee for the EMACs.