Introduce "linux-xilinx" recipe
The linux-xlnx base recipe compiles very very slowly because of all
the config handling. Also we want to undo several of these steps and
that's almost impossible.

Create a linux-xilinx recipe that basically shadows linux-xlnx and make
that the default provider. This also helps picking a version with not
quite such annoying bugs as the 2018.1 has.