partition_sd_card.sh: Avoid "sfdisk -u" and don't attempt to mount
The -uM commandline option is no longer supported by recent sfdisk, so avoid it.
Increase the rootfs size and don't use bigalloc on the data partition.
Don't attempt to mount the partitions, only leads to problems. Just unplug the
card after running the script and put it back to have userspace mount it.