Convert defconfig based kernel into config fragments
Instead of full defconfigs, provide a config fragment for the topic boards. This fixes the issue
that kernel configuration would take longer than actually compiling the kernel. This also gets
rid of the linux-xilinx version that worked around that issue in a different way.

See also:
https://www.yoctoproject.org/docs/2.5/kernel-dev/kernel-dev.html#creating-config-fragments

Steps to convert the configuration:
  bitbake -c kernel_configme -f virtual/kernel
  bitbake -c menuconfig virtual/kernel
Just save the .config file and merge/replace with old defconfig, then run menuconfig
again to validate the changes.
  bitbake -c diffconfig virtual/kernel
This creates the fragment to replace the full defconfig. Only EXT4 got dropped for some reason.

Next step is to convert this to a "kmeta" package to make it a proper BSP