conf: Fix SOC_FAMILY override setting for ultrascale modules
topic-miamimp.inc appended to SOC_FAMILY which caused troubles in scripts using that
particular setting, resulting in a mangled OVERRIDES.

Change the MACHINE_OVERRIDES directly and in correct order, and leave SOC_FAMILY as
is.

This activates the proprietary MALI libraries through SOC_VARIANT="eg", but these don't
actually compile properly in this warrior branch. For the time being, just disable the
libmali-xlnx support so that we can at least continue developing.