devicetree: Make adv6411 chip probe again
Injecting buses into an i2c mux is not possible in a DT overlay, there's no error message but
the new bus doesn't actually work. Register all buses in the "common" devicetree and only add
the chips in the overlay. This makes the adv7611 probe correctly.
Also add GPIO descriptions for the GPIO expanders on the 7-series boards.