bootscript: Increase DMA atomic pool size for zynqmp
The USB3 controller sometimes outputs messages like this:
  xhci-hcd xhci-hcd.0.auto: Ring expansion failed
The cause appears to be that the driver is allocating DMA memory
while in atomic context. This is allocated from a pool whose size
defaults to 256k. Increase the pool size to 512k using a kernel
commandline argument makes the message go away and makes the
wifi dongle work properly.