Fix HDMI input not working on tdkz15 and tdkz30
There's a hidden dependency between the adv7611 driver (adv7604) and the
CEC core. Enable the CEC function of the adv7511 to make the CEC core
become part of the kernel, and then the driver probes. Note that the
config setting for adv7611 CEC support does not actually make it into
the configuration, probably depends on manual selection of drivers.