devicetree: Enable OTG mode for MPSoC devices
To make OTG work, the DWC3 must be in "dual role" and not in "otg"
mode. The USB_DWC3_OTG setting appears to be broken on the MPSoC.

Set the USB controller to OTG mode in the devicetrees. The DWC3
driver apparently has no notion of vbus regulator control, so the
USB supply remains on at all times. The DWC3 controller on the MPSoC
seems to have issues in combination with the SMSC USB3340 PHY on
the GEN board, in peripheral mode it drops to "full-speed", a.k.a.
USB1.1 mode.

Also attach regulators to devices that request them, like IO
expanders.