xdpzu7: Postpone sdhc1 initialization to make WiFi work
Wifi is connected through the FPGA so it cannot be activated until the
FPGA is up. Also properly set up the "pwrseq" driver for the reset.

To be able to boot from SD card, supply an alternative devicetree that
initializes the SD controller early. The install script or boot script
must take this into account.