conf/machine: Set SERIAL_CONSOLES unconditionally
Though bitbake.conf assignes a proper default to SERIAL_CONSOLES, systemd-serialgetty
replaces it with a broken less weak default. To launch getty only on the ports we
allow, assign SERIAL_CONSOLES on all boards to be the same as the kernel console
specified in SERIAL_CONSOLE.

This fixes bluetooth not working on systemd images on the XDP board.