ttpzu9, tdpzu9: Detect and support v1r1 and v1r0 module revisions
kernel: Different GPIO expander on the v1r1, add kernel support and
configure the pcal9554b.

u-boot: Check for the presence of a GPIO expander on address 38. If present,
then this is a v1r0 revision of the module, patch the in-memory devicetree
to change the GPIO expander's reg and compatible properties.

Also, pending kernel support, disable the pull-ups on the pcal9554b.

FPGA: Use board version resistors to select FAN tacho input