u-boot: Fix that some SD cards were not working in SPL
Bug in the MMC driver code in u-boot, turned out Xilinx already had a patch
for it. A similar bug was in the kernel too.

Problem was that the "tap delay" settings ended up being random values instead
of proper defaults.

Most noticable on the tdpzu9 board which has an ultra-speed SD card slot.