kernel: Fix tap delay initialization in Xilinx sdhci driver
The same bug as in the bootloader was in the kernel code too, the code
was assigning data through a macro that'd go out of scope. Now that we
know that, the patches leading up to that error no longer need to be
reverted, we just fix the bug by using normal static variables.