tdpzu9, ttpzu9: Configure TSU clock at 250MHz and fix SyncE 125MHz clock
The GEM2 MAC was configured for a 250MHz timestamp clock, but no clock
was being provided.

Patch the devicetree to provide a 250MHZ TSU clock from the Si5345 chip
and configure the GEM2 driver to use that clock as a timestamping source.

The "magic" bits for the Marvell PHY to enable a 125MHz SyncE output were
wrong, it was outputting a 25MHz clock instead. Set them correctly now.