kernel, clk-si5341: Support NVM programming through sysfs
Add two attributes to the si534x driver to allow the clock to be programmed during
production. One is the output_mask that selectively forces outputs to "off". The
other is the program_nvm_bank that reports the NVM state and programs the NVM when
written with a magic cookie value.