u-boot: tdpzu9, ttpzu9: Add initial delay to allow clock chip to initialize
The Si5345 clock takes a while to initialize and when booting from QSPI the
clock isn't up yet. Attempts to monitor PLL lock status did not work, so just
delay for a few milliseconds when boot mode is QSPI and an external reset
was active.

This fixes that ethernet and PCIe did not work when booting from QSPI. When
booting from SD or eMMC the ROM spends enough time communicating with the
storage device to allow the clock to be up and running when the GTP PLLs
are configured.