tdpzu9: Use pull-down on MIO39 (SD SEL)
Apply a pull-down on MIO pin 39, which is the voltage select for the SD card.
There's already an external pull-down on this pin, and the level shifter also
has an internal pull-down. This is just to prevent that the pull-up on this
pin will oppose the pull-downs, and the pin should remain low by default.

Put this in the devicetree and in the SPL boot code