kernel: Make mdev hotplug work on petalinux
The petalinux distro uses busybox-mdev as device manager, but does not
explicitly configure that anywhere (except for microblaze). Make it
so that the kernel includes mdev hotplug support when the distro did
not set VIRTUAL-RUNTIME_dev_manager to any value.

This makes disks mount automatically and automatically loads kernel
modules when needed as with other distros.