xdp-hwdetect.sh: Use YUYV 4:2:2 format for software capture pipeline
The format for the pad between the CSC and DMA video components was set to a
4:2:0 YUV format, which resulted in the DMA not being capable of YUYV output format
because that requires 4:2:2 video data.

Change the format into UYVY8_1X16 to make the DMA support YUYV formats. This enables
the PR demo application to use the software image pipeline again.

The VCU demo also remains functional after this modification.