xdp-hwdetect: Fix that the Dyplo IO camera shows only top-left corner
The camera defaults to 4k video mode, and the pipeline behind it just crops
away the top left corner since it's configured for 1080p.

The name of the IMX camera in the initialization script was wrong. Provide
the proper name in the script and then the camera switches to 1080p.

This solves that in the PR demo whatever the camera was directly pointing at
would appear in the bottom-right corner of the produced video. This also
doubles the field of view.