u-boot: tdkz: Fix that 1GB boards would no longer boot
The patch "0001-UBI-support-u-boot.patch" contained several issues:
- Placed outside the re-include guard macro
- Redefines the boot variables instead of only patching what's needed
- Spelling error 'quite' in KERNELARGS which was unused anyway

Properly change the qspi boot mode to use ubi, and remove KERNELARGS

Preserve the fdt_high environment, thus enabling 1GB boards to boot properly. Without
this setting, u-boot will place the devicetree at the end of RAM, close to the 1GB
limit, and this causes the kernel to crash when it parses the devicetree.

Fixes: 45f2154ed ("Reduce number of SPI to two")