u-boot: tdkz, tdkzu: Remove network support and environment storage
On the tdkz and tdkzu boards the ethernet is connected through logic, is not
configured by petalinux, and thus isn't available in u-boot. Remove network
support to reduce the size and avoid a boot error message.

Storing the full environment in SPI causes issues when upgrading the bootloader,
wastes time at boot and displays an error message. Remove the support to get
rid of the error message and speed up booting.