tdkz: Correct QSPI-NOR partitions and make UBIFS work
The rootfs did not fill the flash remainder but left a "gap" at the end. This
usually results in problems with ubifs in u-boot, so fix that.

The newer modules have a 64MB NOR flash chip, so increase the partition to
match that. If the detected flash is smaller, the kernel will "truncate" it,
so the 32MB modules will also keep functioning normally.

Change the u-boot "mtdparts" setting to automatically detect and use the flash
size.

Disable 4K sectors in u-boot, to match the kernel sector size and to be able
to use UBI at all, which requires at least 16k sectors.