kernel: Fix that QSPI NOR failed to read in quad mode on newer flash devices
Modern SPI flash chips have extra tables to tell the driver what instructions
to use. This takes precedence over the "manual" settings.

The Zynq QSPI controller is uncapable of 4-byte addres modes, resulting in
corrupted reads on devices that provide a 4-Byte Address Instruction Table.
Prevent this by not parsing the 4BAIT table if the controller is a Zynq QSPI.

This fixes that the new 64MB NOR flash chips could not be read by Linux.