fpga-bit-to-bin: Use python from PATH, don't patch the hashbang
When the script is taken from a sstate-cache object built elsewhere, patching
the hashbang results in failures like this:
  ... bad interpreter: No such file or directory

Instead of /usr/bin/python, use "/usr/bin/env python" as interpreter, which
will locate Python on the PATH and thus use the bitbake supplied one as
intended.

This also makes the fpga-bit-to-bin independent of the build environment.