kernel: Add support for Linux' own USB network adapter
Add the USB CDC drivers so that the boards can connect to themselves or
another board that is exposing a Linux USB network gadget.

This allows to plug the USB host of the xdpzu7 into the USB peripheral of
a tdpzu9 for example and get a working ethernet link between the two.