kernel-module-88x2cu: Add bluetooth support
Add support for the bluetooth part of the 8822cu chipset. Petalinux is using an older
version of OE and hence an older version of linux-firmware, so the firmware files do not
always match the ones required in the kernel.

Create a recipe that installs a new version for the bluetooth part of the 8822cu chipset,
this fixes that the bluetooth part would not function.