QSPI partitioning cleanup
Use the same partitioning for all boards, that is:
* For 7-series, use 1MB for boot and the remainder for rootfs
* For Ultrascale, use 1.5MB for boot and remainder for rootfs
So on all boards, there are only 2 QSPI partitions now. This allows the scripts for
updating the QSPI data to be independent of the SoM.

This puts the rootfs in the same location as the non-petalinux BSP, making it possible
to switch between the two without having to completely reformat the NOR flash.

Fix the boot arguments in the boot scripts, and refer to the rootfs partition by name
instead of by number.

Change the "compatible" string for the NOR flash into "jedec,spi-nor". This gets rid
of the warnings at boot about the flash type not matching the devicetree.

Update the "mtdparts" variable in u-boot to match the common partitioning scheme.

Also ensure that u-boot doesn't waste time attempting to load the environment from SPI
by setting ENV_IS_NOWHERE in the configuration.