device-tree: Move USB resets to SOM, add supplies to ttpzu9
The USB PHY resets are part of the SOM itself, regardless of whether they
are being used, so move them to zynqmp-topic-miamiplusmp.dts.

Remove the Xilinx "reset" through mode pins (again), the property name
is now properly 'reset-gpios' so remove that too.

Add the power supply settings to the ttpzu9 as well.