devicetree, miamiplusmp: Assign all MAC addresses in EEPROM
The production test writes 4 MAC addresses in to the EEPROM on the
SOM. Two in the "protected" area, assigned by the operator/software
and two more at the start of the EEPROM that are semi-random.

Assign these addresses to all four MAC devices (gem0 .. gem3) as
a default. Projects can override these if they wish.