tspzu: Configure memory size automatically
Read DDR setting registers in u-boot and adjust the memory bank sizes
based on that. Pass the memory information on to the Linux OS.
Set the devicetree to have the maximum possible size (8GB) by default.

This only applies to the tspzu modules.

Tested with a module with only 2GB of 32-bit memory, and a standard 4GB
module.