linux-xlnx: Fix xilinx_phy resume and pcie_xilinx startup
Fix that the Xilinx PCS/PMA PHY could not resume properly (i.e. any down-up
sequency would break it).

Fix the reset/init routine of the pcie-xilinx core so that it supports
a GPIO reset signal and waits for the link to become active. This fixes
that the core usually wouldn't work on a 7-series device.