kernel: Retrieve xtal input frequency from EEPROM
This replaces the method of using a GPIO to select the frequency. Instead,
the clock frequency is read from EEPROM contents.

Add a shell script to quickly program the EEPROM in existing boards. This
should be done before upgrading the software, as it will need access to
the clock setting. The script can be copied into a terminal and doesn't
need to be deployed on the board.