kernel: Work around OTG enumeration issues on the tdpzu
Plug an OTG device into the tdpzu, this results in:
  usb 1-1: Dual-Role OTG device on non-HNP port
  usb 1-1: set a_alt_hnp_support failed: -32
  usb usb1-port1: unable to enumerate USB device
Adding a USB hub in between solves this.

Modify the hub code to not give up immediately when OTG negotiations
run into trouble, but just continue as a standard hub, which lets the
cable decide which end will be what.