fsbl: On tspzu only override defaults when EEPROM values are valid
Prepare to have different defaults depending on machine settings.
Thus only "force" the density/width/ecc setting if the EEPROM setting
makes sense, and use the default otherwise.

This is important for production testing a SOM, as the memory config
isn't programmed yet at first boot, and needs to be provided externally.