machine: Add "powerkey" feature to boards that have a power button
Most boards have a power button that turns the board on and off. To
power down the board gracefully on pressing that button, some userspace
support is needed (systemd provides this, with initscripts there is no
such built-in support). Add a MACHINE_FEATURE "powerkey" to indicate
that the button is there.