fsbl-firmware, tspzu: Clear full memory range when ECC enabled
When ECC is enabled, the FSBL must clear (write to) the full range
of DDR memory. The generated FSBL only cleared the first 4GB, it
did not take the EEPROM DDR settings into account.
Calculate the memory size to be cleared using the DDR registers.

This fixes that the boards with 8GB of ECC DDR memory (72-bits bus)
would hang while booting into the kernel.