tspzu, tdkzu: Activate lima driver for MPSoC modules
Use MALI400 on the ADV7511 HDMI output
Add "xlnx" and "axi_hdmi_drm" entry to mesa.
Configure X server to use the MALI as accellerator.

The tspzu modules have a MALI400, activate the corresponding entries
in the xorg.conf file to make that work. Leave commented-out lines
in place that show how to avoid using the GPU.