tepzu-uboot.dts: Deactivate USB HUB to avoid boot failure
The DWC3 driver in U-boot cannot cope with the USB HUB entries in the
devicetree, resulting in a boot failure:
  Failed to bind: , err: -2
  initcall failed at call 000000007fef2c60 (err=-2)
The driver assumes that the first child node of the USB controller is
the DWC3 part, but gets the hub instead. Since the HUB won't work anyway
until it's been properly reset in Linux, just disable the controller.