xserver-xf86-config: Workaround ZynqMP DisplayPort crashes
Using the ZynqMP DisplayPort output, attempting to use more than 16-bit
mode, OpenGL or a hardware cursor will crash the Xserver. This appears to
be some issue with the ZynqMP DP driver. However, GLES works, so if we
disable the "glx" extension and enable the SWCursor option, we get an
accelerated X11 server and we can run GLES applications.

Make "disable glx" and "enable SWCursor" the default for tspzu boards.