kernel: Increase CONFIG_USB_GADGET_STORAGE_NUM_BUFFERS
When using a mass_storage gadget over USB, the "num_buffers" parameter
defaults to "2". Each buffer is 16kB in size. This is hampering performance
considerably. Some test results:
  num_buffers=2,  from NVME:  186MB/s
  num_buffers=2, from cache:  224MB/s
  num_buffers=8,  from NVME:  290MB/s
  num_buffers=8, from cache:  348MB/s
Results obtained from a tdkzu9 board using one of the USB-C ports connected
to a PC. Measured using this command on the PC host:
  sudo dd if=/dev/sdd of=/dev/null bs=1M count=2560 iflag=direct
The board shares its /dev/nvme0n1 device as read-only mass storage. The
cache result is obtained by simply repeating the dd command, which will
retrieve the data from the cache.

Set the default to the higher values. On the 7-series with only USB2
capability, increase to "4", on the MPSoC boards set it to "8".